Document some internal `fxa-*` claims inside JWT access tokens.
Inherited from BrowserID, some internal Mozilla reliers need access
to special `fxa-*` prefixed claims. These were only documented in
the part of the API doc about BrowserID, so I've copied the text
over to the OAuth docs to make it easier to find, and added a few
notes about how we expect the claims to be used.